LayoutTests/media/media-source/media-source-duration-after-append-expected.txt


EVENT(sourceopen)
RUN(sourceBuffer = source.addSourceBuffer("video/mock; codecs=mock"))
RUN(sourceBuffer.appendBuffer(initSegment))
EVENT(updateend)
EXPECTED (video.duration == '0') OK
RUN(sourceBuffer.appendBuffer(sample))
EVENT(updateend)
EXPECTED (video.duration == '10') OK
END OF TEST
